requirements.txt
django
django-dotenv
psycopg2-binary
gunicorn
django-environ
psycopg2
whitenoise